# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UpPay Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerList"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /v1/customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /v1/customers/{id}/payments:
    get:
      operationId: getCustomerPayments
      tags: [customers]
      summary: Get payments for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

  # ─── Integrated Systems ──────────────────────────────────────────────────────
...
  /v1/api-keys:
    get:
      operationId: listApiKeys
      tags: [api-keys]
      summary: List API keys for a system
      parameters:
        - name: system_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [api-keys]
      summary: Create a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyWithSecret"
  /v1/api-keys/{id}/revoke:
    post:
      operationId: revokeApiKey
      tags: [api-keys]
      summary: Revoke an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"

  # ─── Subscriptions ───────────────────────────────────────────────────────────
...
  /v1/fees:
    get:
      operationId: listFees
      tags: [fees]
      summary: List provider fee configurations
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderFee"
    post:
      operationId: createFee
      tags: [fees]
      summary: Create a fee configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderFeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderFee"
  /v1/fees/{id}:
    patch:
      operationId: updateFee
      tags: [fees]
      summary: Update a fee configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderFeeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderFee"

  # ─── Audit Logs ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Payment:
      type: object
      required:
        - id
        - amount
        - currency
        - payment_method
        - status
        - provider
        - source_system
        - created_at
      properties:
        id:
          type: string
        source_system_id:
          type: ["string", "null"]
        customer_id:
          type: ["string", "null"]
        provider:
          type: string
        provider_payment_id:
          type: ["string", "null"]
        status:
          type: string
        external_reference:
          type: ["string", "null"]
        amount:
          type: integer
        gross_amount:
          type: ["integer", "null"]
        provider_fee:
          type: ["integer", "null"]
        net_amount:
          type: ["integer", "null"]
        currency:
          type: string
        payment_method:
          type: string
        description:
          type: ["string", "null"]
        installments:
          type: ["integer", "null"]
        qr_code:
          type: ["string", "null"]
        qr_code_base64:
          type: ["string", "null"]
        copy_paste:
          type: ["string", "null"]
        boleto_url:
          type: ["string", "null"]
        boleto_barcode:
          type: ["string", "null"]
        boleto_digitable_line:
          type: ["string", "null"]
        payment_link_url:
          type: ["string", "null"]
        source_system:
          type: string
        callback_url:
          type: ["string", "null"]
        expires_at:
          type: ["string", "null"]
        due_date:
          type: ["string", "null"]
        paid_at:
          type: ["string", "null"]
        cancelled_at:
          type: ["string", "null"]
        refunded_at:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
...
    Customer:
      type: object
      required: [id, name, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        document:
          type: ["string", "null"]
        document_type:
          type: ["string", "null"]
        address_city:
          type: ["string", "null"]
        address_state:
          type: ["string", "null"]
        asaas_customer_id:
          type: ["string", "null"]
        pagbank_customer_id:
          type: ["string", "null"]
        created_by_system_id:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
    CustomerList:
      type: object
      required: [data, total, limit, offset]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    CustomerAddress:
      type: object
      properties:
        street:
          type: string
        number:
          type: string
        district:
          type: string
        city:
          type: string
        state:
          type: string
        zip_code:
          type: string
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        document:
          type: string
        document_type:
          type: string
        address:
          $ref: "#/components/schemas/CustomerAddress"

    # ─── Integrated System ──────────────────────────────────────────────────────
...
    ApiKey:
      type: object
      required: [id, system_id, name, environment, status, created_at]
      properties:
        id:
          type: string
        system_id:
          type: string
        name:
          type: string
        key_prefix:
          type: string
        environment:
          type: string
        status:
          type: string
        permissions:
          type: array
          items:
            type: string
        last_used_at:
          type: ["string", "null"]
        created_at:
          type: string
    ApiKeyWithSecret:
      type: object
      required: [id, system_id, name, environment, status, key, created_at]
      properties:
        id:
          type: string
        system_id:
          type: string
        name:
          type: string
        key_prefix:
          type: string
        key:
          type: string
        environment:
          type: string
        status:
          type: string
        permissions:
          type: array
          items:
            type: string
        created_at:
          type: string
    ApiKeyInput:
      type: object
      required: [system_id, name, environment]
      properties:
        system_id:
          type: string
        name:
          type: string
        environment:
          type: string
        permissions:
          type: array
          items:
            type: string

    # ─── Subscription ────────────────────────────────────────────────────────────
...
    ProviderFee:
      type: object
      required: [id, provider, payment_method, fee_type, active, created_at]
      properties:
        id:
          type: string
        provider:
          type: string
        payment_method:
          type: string
        fee_type:
          type: string
        fixed_amount:
          type: ["integer", "null"]
        percentage_rate:
          type: ["number", "null"]
        settlement_days:
          type: ["integer", "null"]
        active:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
    ProviderFeeInput:
      type: object
      required: [provider, payment_method, fee_type]
      properties:
        provider:
          type: string
        payment_method:
          type: string
        fee_type:
          type: string
        fixed_amount:
          type: integer
        percentage_rate:
          type: number
        settlement_days:
          type: integer
    ProviderFeeUpdate:
      type: object
      properties:
        fixed_amount:
          type: integer
        percentage_rate:
          type: number
        settlement_days:
          type: integer
        active:
          type: boolean

    # ─── Audit Log ───────────────────────────────────────────────────────────────
...